update dropdown filter schema and layout
